[Interactive Graph Editor] Moving around/renaming components
