+++ .changeset/mean-needles-float.md
@@ -1,0 +1,6 @@
+---
+"@khanacademy/perseus": patch
+"@khanacademy/perseus-editor": patch
+---
+
+[Interactive Graph Editor] Moving around/renaming components so they make more sense for the upcoming hint mode work
