+++ packages/perseus-editor/src/components/coordinate-pair-input.tsx
