+++ packages/perseus-editor/src/components/interactive-graph-settings.tsx
@@ -12,13 +12,12 @@ import Banner from "@khanacademy/wonder-blocks-banner";
 import {View} from "@khanacademy/wonder-blocks-core";
 import {OptionItem, SingleSelect} from "@khanacademy/wonder-blocks-dropdown";
 import {color, spacing} from "@khanacademy/wonder-blocks-tokens";
-import {LabelSmall} from "@khanacademy/wonder-blocks-typography";
 import {css, StyleSheet} from "aphrodite";
 import * as React from "react";
 import _ from "underscore";
 
 import LabeledRow from "./graph-locked-figures/labeled-row";
-import ToggleableCaret from "./toggleable-caret";
+import Heading from "./heading";
 
 import type {PerseusImageBackground} from "@khanacademy/perseus";
 
@@ -32,38 +31,6 @@ const defaultBackgroundImage = {
 
 function numSteps(range: any, step: any) {
     return Math.floor((range[1] - range[0]) / step);
-}
-
-function Heading({
-    title,
-    isOpen,
-    onToggle,
-}: {
-    title: string;
-    isOpen: boolean;
-    onToggle?: (isOpen: boolean) => void;
-}) {
-    return (
-        <View
-            style={{
-                flexDirection: "row",
-                justifyContent: "space-between",
-                backgroundColor: color.offBlack8,
-                padding: spacing.xSmall_8,
-                marginTop: spacing.small_12,
-                // NOTE(jeremy): This is the inverse of the @editorPadding CSS
-                // variable found in perseus-editor.less. For now, it must
-                // match otherwise there's a gap from this header to the edge
-                // of the editor borders.
-                marginInline: -10,
-                cursor: "pointer",
-            }}
-            onClick={() => onToggle?.(!isOpen)}
-        >
-            <LabelSmall>{title}</LabelSmall>
-            <ToggleableCaret isExpanded={isOpen} />
-        </View>
-    );
 }
 
 type Range = [min: number, max: number];
