+++ packages/perseus/src/index.ts
@@ -173,6 +173,7 @@ export type {
     // Utility types
     Range,
     Size,
+    CollinearTuple,
     MathFormat,
     InputNumberWidget, // TODO(jeremy): remove?
     // Widget configuration types
