+++ packages/perseus/src/styles/perseus-renderer.less
@@ -718,7 +718,7 @@
     }
 }
 
-.perseus-widget-editor-content .locked-figure-accordion h2 {
+.perseus-widget-editor-content .perseus-editor-accordion h2 {
     // Reset the padding that's added to article h2
     // for the accordion h2.
     padding-top: 0;
